Added Login Via API and added, extra fields to some migrations and added suppliers table

+++ src/database/migrations/2021_02_21_203841_create_products_table.php
@@ -23,7 +23,7 @@ class CreateProductsTable extends Migration
             $table->bigInteger('item_code');
             $table->string('description');
             $table->integer('quantity');
-            $table->double('price', 8, 2);
+            $table->decimal('price', 10, 2);
             $table->timestamps();
             $table->softDeletes();
         });
